FIX: Boundary condition implementation for benchmark Stokes_MOR. Changes include: 1. Remove all boundary conditions for pressure, but pin a reference value for pressure in the cell at the bottom left corner; 2. Evaluate values for velocity boundary conditions by using evaluate_CornerFlow_MOR() on the actual boundaries instead of the interior boundary points. It is related to the change of Dirichlet implementation ; 3. Loop k from 1 to n_bc-2 instead of from 0 to n_bc-1 to assign boundary conditions for the interior points associated with the up and down boundaries. Otherwise the starting and ending points are overrided with a false value because of the change No.2.
After the changes, the numerical results for velocities match quite well with the analytical results. But the contours of pressures differ from one to the other, especially for small pressure. If giving boundary condition for pressure on the whole bottom boundary, the problem is solved that numerical results match with analytical ones for velocites and pressure. But WHY do we need more boundary conditions than necessary?