FD-PDE Enthalpy: Refactor and fix some bugs by adding smaller tests. This is a bigger commit, so I will summarise:
1. Tests:
tests/test_enthalpy_2d_diffusion.c -> 2d analytical diffusion to test the fdpde implementation with 2 variables (H/TP,C) but also multiple components. This test runs for H and 3 chemical components (3 dofs in total), and solves the same diffusion equation for each unknown. Also, it tests the enthalpy method function pointer in a simple way because in order to match the equations, an artificial enthalpy method needed to be set up. The convergence error is the same for all variables.
This is a good test also for ierr = FDPDEEnthalpyUpdateDiagnostics(fd,dm,x,&dmnew,&xnew); used for output all enthalpy variables.

tests/test_enthalpy_1d_solidification_TC.c -> fixed the 1d solidification problem with TC as primary variables (i.e. Krabbenhoft et al). This was a good problem playing with the initial condition and time-stepping (which is very sensitive, as we discussed together). An interesting observation: for the default parameters, when the solidification front slows down after a couple of timesteps, SNES goes into a cycle (2 fast-solve timesteps with 1-2 SNES iterations followed by 1 slow-solve with many iterations ~30 SNES iterations) which is reflected into the solution (during fast solves, the solidification front hasn't evolved much, while during the fast-solve the front evolved more). This creates the step-pattern in the solution for the solidification front and temperature evolution. Note that there is a fixed timestep size during a simulation as it is a diffusion-type problem (potentially one can evolve it with the solidification front velocity, but didn't bother).

tests/test_enthalpy_1d_eutectic_solidification.c -> fixed the 1d solidification problem with an eutectic phase diagram (HC). I used the formulation and reproduced the results from Parkinson et al (2020), and modified the analytical solution from Worster (1992) to be compatible with the non-dimensionalization. After fixing some bugs with the previous tests, the trick for good convergence was to use a small timestep size, so I introduced a CFL criterion: dt = cfl*dx/v, where v is the frame velocity. The final thing was that curly(C) in Parkinson et al and Worster, which is an input parameter, is not the same. So in this implementation and notes they have a different symbol.

To execute the full benchmarks:
python test_enthalpy_2d_diffusion.py
python test_enthalpy_1d_solidification_TC.py
python test_enthalpy_1d_eutectic_solidification.py

2. The Enthalpy Method:
- I removed the extra coefficients (a-e) related to the enthalpy method in the previous commit, and only pass the the enthalpy method as a single pointer (black box) with:
	ierr = FDPDEEnthalpySetEnthalpyMethod(fd,Form_Enthalpy,enthalpy_method_description,usr);
and the user has to define a function of the form:
	Form_Enthalpy(fd,i,j,H,C[],&P,&TP,&T,&phi,CS,CF,usr);*
where the variables inside mean:
    fd - FDPDE object
    i  - cell index (1-dir)
    j  - cell index (2-dir)
    H  - enthalpy (input if H-primary variable)
    C[]- bulk composition (input) (ncomp-1)
    P  - pressure - needs to be specified by the user but can be easily calculated inside with (i,j)indices/usr for lithostatic pressure
    TP - primary temperature variable (input if TP-primary variable)
    T  - secondary temperature variable
    phi- porosity
    CS[] - solid composition (ncomp-1)
    CF[] - fluid composition (ncomp-1)
    ncomp - number of components
    user - user context provided

*please see the example tests for details on how it works!

3. Bugs fixes in this commit:
- Line 260 in fdpde_enthalpy_residual.c (index)
- test-related (see above)