DMStagOutput Read: Added reverse functionality to read binary file that was written with DMStagViewBinaryPython()
Description:
If a dm/x is written to file with:
	ierr = DMStagViewBinaryPython(dm,x,fname);CHKERRQ(ierr);
Then it can be read back with:
	ierr = DMStagReadBinaryPython(&dm,&x,fname);CHKERRQ(ierr);

Warning: After reading, dm and x need to be destroyed! Implemented only sequentially and for 2-d DMStag, but could easily be extended.

Test:
tests/test_dmstagoutput_read.c - it creates a dm/x with various dofs and fields, writes it to file and then reads it.

Run test:
python test_dmstagoutput_read.py - checks the norm between the written and read vectors (should be 0 for all combination of dofs and coordinates).

Optimization:
Is there a more elegant way to do lines 941-969 in dmstagoutput.c? i.e., put back vector components extracted with DMStagVecSplitToDMDA.